GraqlController titan random failing test fix (#1601)
* Split GraqlController tests into three classes

* Fix randomly failing POST and DELETE test
