+++ grakn-test/src/test/java/ai/grakn/test/engine/controller/GraqlControllerGETTest.java
@@ -30,6 +30,7 @@ import ai.grakn.test.engine.controller.TasksControllerTest.JsonMapper;
 import ai.grakn.util.REST;
 import com.jayway.restassured.RestAssured;
 import com.jayway.restassured.response.Response;
+import java.util.Collections;
 import mjson.Json;
 import org.junit.AfterClass;
 import org.junit.Before;
@@ -37,10 +38,9 @@ import org.junit.BeforeClass;
 import org.junit.ClassRule;
 import org.junit.FixMethodOrder;
 import org.junit.Test;
+
 import org.junit.runners.MethodSorters;
 import spark.Service;
-
-import java.util.Collections;
 
 import static ai.grakn.engine.GraknEngineServer.configureSpark;
 import static ai.grakn.graql.internal.hal.HALBuilder.renderHALArrayData;
@@ -53,10 +53,10 @@ import static ai.grakn.util.REST.Request.Graql.LIMIT_EMBEDDED;
 import static ai.grakn.util.REST.Request.Graql.MATERIALISE;
 import static ai.grakn.util.REST.Request.Graql.QUERY;
 import static ai.grakn.util.REST.Request.KEYSPACE;
-import static ai.grakn.util.REST.Response.ContentType.APPLICATION_HAL;
 import static ai.grakn.util.REST.Response.ContentType.APPLICATION_JSON;
 import static ai.grakn.util.REST.Response.ContentType.APPLICATION_JSON_GRAQL;
 import static ai.grakn.util.REST.Response.ContentType.APPLICATION_TEXT;
+import static ai.grakn.util.REST.Response.ContentType.APPLICATION_HAL;
 import static ai.grakn.util.REST.Response.EXCEPTION;
 import static ai.grakn.util.REST.Response.Graql.ORIGINAL_QUERY;
 import static ai.grakn.util.REST.Response.Graql.RESPONSE;
@@ -73,6 +73,7 @@ import static org.junit.Assume.assumeTrue;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyBoolean;
 import static org.mockito.ArgumentMatchers.argThat;
+import static org.mockito.ArgumentMatchers.booleanThat;
 import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.RETURNS_DEEP_STUBS;
 import static org.mockito.Mockito.doAnswer;
@@ -83,7 +84,7 @@ import static org.mockito.Mockito.when;
 
 //TODO Run in name order until TP Bug #13730 Fixed
 @FixMethodOrder(MethodSorters.NAME_ASCENDING)
-public class GraqlControllerTest {
+public class GraqlControllerGETTest {
 
     private static final String HOST = "localhost";
     private static final int PORT = 4567;
@@ -135,10 +136,9 @@ public class GraqlControllerTest {
 
     @Test
     public void GETGraqlMatch_QueryIsExecuted(){
-        String query = "match $x isa person;";
+        String query = "match $x isa movie;";
         sendGET(query, APPLICATION_TEXT);
 
-        //noinspection ResultOfMethodCallIgnored
         verify(mockGraph.graql().materialise(anyBoolean()).infer(anyBoolean()))
                 .parse(argThat(argument -> argument.equals(query)));
     }
@@ -204,7 +204,7 @@ public class GraqlControllerTest {
     @Test
     public void GETGraqlMatchNoMaterialise_ResponseStatusIs400(){
         Response response = RestAssured.with().queryParam(KEYSPACE, mockGraph.getKeyspace())
-                .queryParam(QUERY, "match $x isa person;")
+                .queryParam(QUERY, "match $x isa movie;")
                 .queryParam(INFER, true)
                 .accept(APPLICATION_TEXT)
                 .get(String.format("http://%s:%s%s", HOST, PORT, REST.WebPath.Graph.GRAQL));
@@ -215,24 +215,22 @@ public class GraqlControllerTest {
 
     @Test
     public void GETGraqlMatchWithReasonerTrue_ReasonerIsOnWhenExecuting(){
-        sendGET("match $x isa person;", APPLICATION_TEXT, true, true, 0);
+        sendGET("match $x isa movie;", APPLICATION_TEXT, true, true, 0);
 
-        //noinspection ResultOfMethodCallIgnored
-        verify(mockQueryBuilder).infer(true);
+        verify(mockQueryBuilder).infer(booleanThat(arg -> arg));
     }
 
     @Test
     public void GETGraqlMatchWithReasonerFalse_ReasonerIsOffWhenExecuting(){
-        sendGET("match $x isa person;", APPLICATION_TEXT, false, true, 0);
+        sendGET("match $x isa movie;", APPLICATION_TEXT, false, true, 0);
 
-        //noinspection ResultOfMethodCallIgnored
-        verify(mockQueryBuilder).infer(false);
+        verify(mockQueryBuilder).infer(booleanThat(arg -> !arg));
     }
 
     @Test
     public void GETGraqlMatchWithNoInfer_ResponseStatusIs400(){
         Response response = RestAssured.with().queryParam(KEYSPACE, mockGraph.getKeyspace())
-                .queryParam(QUERY, "match $x isa person;")
+                .queryParam(QUERY, "match $x isa movie;")
                 .accept(APPLICATION_TEXT)
                 .get(String.format("http://%s:%s%s", HOST, PORT, REST.WebPath.Graph.GRAQL));
 
@@ -242,24 +240,22 @@ public class GraqlControllerTest {
 
     @Test
     public void GETGraqlMatchWithMaterialiseFalse_MaterialiseIsOffWhenExecuting(){
-        sendGET("match $x isa person;", APPLICATION_TEXT, false, false, 0);
+        sendGET("match $x isa movie;", APPLICATION_TEXT, false, false, 0);
 
-        //noinspection ResultOfMethodCallIgnored
-        verify(mockQueryBuilder).materialise(false);
+        verify(mockQueryBuilder).materialise(booleanThat(arg -> !arg));
     }
 
     @Test
     public void GETGraqlMatchWithMaterialiseTrue_MaterialiseIsOnWhenExecuting(){
-        sendGET("match $x isa person;", APPLICATION_TEXT, false, true, 0);
+        sendGET("match $x isa movie;", APPLICATION_TEXT, false, true, 0);
 
-        //noinspection ResultOfMethodCallIgnored
-        verify(mockQueryBuilder).materialise(true);
+        verify(mockQueryBuilder).materialise(booleanThat(arg -> arg));
     }
 
     @Test
     public void GETGraqlMatchWithHALTypeAndNumberEmbedded1_ResponsesContainAtMost1Concept(){
         Response response =
-                sendGET("match $x isa person;", APPLICATION_HAL, false, true,1);
+                sendGET("match $x isa movie;", APPLICATION_HAL, false, true,1);
 
         jsonResponse(response).asJsonList().forEach(e -> {
              assertThat(e.asJsonMap().get("_embedded").asJsonMap().size(), lessThanOrEqualTo(1));
@@ -292,7 +288,7 @@ public class GraqlControllerTest {
 
     @Test
     public void GETGraqlMatchWithHALType_ResponseContainsOriginalQuery(){
-        String query = "match $x isa person;";
+        String query = "match $x isa movie;";
         Response response = sendGET(query, APPLICATION_HAL);
 
         assertThat(originalQuery(response), equalTo(query));
@@ -310,12 +306,12 @@ public class GraqlControllerTest {
         Response response = sendGET(APPLICATION_TEXT);
 
         assertThat(stringResponse(response).length(), greaterThan(0));
-        assertThat(stringResponse(response), stringContainsInOrder(Collections.nCopies(10, "isa person")));
+        assertThat(stringResponse(response), stringContainsInOrder(Collections.nCopies(10, "isa movie")));
     }
 
     @Test
     public void GETGraqlMatchWithTextType_ResponseContainsOriginalQuery(){
-        String query = "match $x isa person;";
+        String query = "match $x isa movie;";
         Response response = sendGET(APPLICATION_TEXT);
 
         assertThat(originalQuery(response), equalTo(query));
@@ -337,7 +333,7 @@ public class GraqlControllerTest {
 
     @Test
     public void GETGraqlMatchWithGraqlJsonType_ResponseIsCorrectGraql(){
-        String query = "match $x isa person;";
+        String query = "match $x isa movie;";
         Response response = sendGET(APPLICATION_JSON_GRAQL);
 
         Json expectedResponse = Json.read(
@@ -347,7 +343,7 @@ public class GraqlControllerTest {
 
     @Test
     public void GETGraqlMatchWithGraqlJsonType_ResponseContainsOriginalQuery(){
-        String query = "match $x isa person;";
+        String query = "match $x isa movie;";
         Response response = sendGET(APPLICATION_JSON_GRAQL);
 
         assertThat(originalQuery(response), equalTo(query));
@@ -362,7 +358,7 @@ public class GraqlControllerTest {
 
     @Test
     public void GETGraqlInsert_ResponseExceptionContainsReadOnlyAllowedMessage(){
-        String query = "insert $x isa person;";
+        String query = "insert $x isa movie;";
         Response response = sendGET(query, APPLICATION_TEXT);
 
         assertThat(exception(response), containsString("read-only"));
@@ -370,7 +366,7 @@ public class GraqlControllerTest {
 
     @Test
     public void GETGraqlInsert_ResponseStatusCodeIs405NotSupported(){
-        String query = "insert $x isa person;";
+        String query = "insert $x isa movie;";
         Response response = sendGET(query, APPLICATION_TEXT);
 
         assertThat(response.statusCode(), equalTo(405));
@@ -378,7 +374,7 @@ public class GraqlControllerTest {
 
     @Test
     public void GETGraqlAggregateWithHalType_ResponseStatusCodeIs406(){
-        String query = "match $x isa person; aggregate count;";
+        String query = "match $x isa movie; aggregate count;";
         Response response = sendGET(query, APPLICATION_HAL);
 
         assertThat(response.statusCode(), equalTo(406));
@@ -386,7 +382,7 @@ public class GraqlControllerTest {
 
     @Test
     public void GETGraqlAggregateWithTextType_ResponseStatusIs200(){
-        String query = "match $x isa person; aggregate count;";
+        String query = "match $x isa movie; aggregate count;";
         Response response = sendGET(query, APPLICATION_TEXT);
 
         assertThat(response.statusCode(), equalTo(200));
@@ -394,19 +390,19 @@ public class GraqlControllerTest {
 
     @Test
     public void GETGraqlAggregateWithTextType_ResponseIsCorrect(){
-        String query = "match $x isa person; aggregate count;";
+        String query = "match $x isa movie; aggregate count;";
         Response response = sendGET(query, APPLICATION_TEXT);
 
         // refresh graph
         graphContext.graph().close();
 
-        int numberPeople = graphContext.graph().getEntityType("person").instances().size();
+        int numberPeople = graphContext.graph().getEntityType("movie").instances().size();
         assertThat(stringResponse(response), equalTo(Integer.toString(numberPeople)));
     }
 
     @Test
     public void GETGraqlAggregateWithTextType_ResponseContentTypeIsText(){
-        String query = "match $x isa person; aggregate count;";
+        String query = "match $x isa movie; aggregate count;";
         Response response = sendGET(query, APPLICATION_TEXT);
 
         assertThat(response.contentType(), equalTo(APPLICATION_TEXT));
@@ -414,7 +410,7 @@ public class GraqlControllerTest {
 
     @Test
     public void GETGraqlCompute_ResponseContainsOriginalQuery(){
-        String query = "compute count in person;";
+        String query = "compute count in movie;";
         Response response = sendGET(query, APPLICATION_TEXT);
 
         assertThat(originalQuery(response), equalTo(query));
@@ -422,7 +418,7 @@ public class GraqlControllerTest {
 
     @Test
     public void GETGraqlComputeWithTextType_ResponseContentTypeIsText(){
-        String query = "compute count in person;";
+        String query = "compute count in movie;";
         Response response = sendGET(query, APPLICATION_TEXT);
 
         assertThat(response.contentType(), equalTo(APPLICATION_TEXT));
@@ -430,7 +426,7 @@ public class GraqlControllerTest {
 
     @Test
     public void GETGraqlComputeWithTextType_ResponseStatusIs200(){
-        String query = "compute count in person;";
+        String query = "compute count in movie;";
         Response response = sendGET(query, APPLICATION_TEXT);
 
         assertThat(response.statusCode(), equalTo(200));
@@ -438,16 +434,16 @@ public class GraqlControllerTest {
 
     @Test
     public void GETGraqlComputeWithTextType_ResponseIsCorrect(){
-        String query = "compute count in person;";
+        String query = "compute count in movie;";
         Response response = sendGET(query, APPLICATION_TEXT);
 
-        int numberPeople = graphContext.graph().getEntityType("person").instances().size();
+        int numberPeople = graphContext.graph().getEntityType("movie").instances().size();
         assertThat(stringResponse(response), equalTo(Integer.toString(numberPeople)));
     }
 
     @Test
     public void GETGraqlComputeWithHALType_ResponseStatusIs406(){
-        String query = "compute count in person;";
+        String query = "compute count in movie;";
         Response response = sendGET(query, APPLICATION_HAL);
 
         assertThat(response.statusCode(), equalTo(406));
@@ -460,13 +456,13 @@ public class GraqlControllerTest {
         assumeTrue(usingTitan());
 
         String fromId = graphContext.graph().getResourcesByValue("The Muppets").iterator().next().owner().getId().getValue();
-        String toId = graphContext.graph().getResourcesByValue("Kermit The Frog").iterator().next().owner().getId().getValue();
+        String toId = graphContext.graph().getResourcesByValue("comedy").iterator().next().owner().getId().getValue();
 
         String query = String.format("compute path from \"%s\" to \"%s\";", fromId, toId);
         Response response = sendGET(query, APPLICATION_TEXT);
 
         assertThat(response.statusCode(), equalTo(200));
-        assertThat(stringResponse(response), containsString("isa has-cast"));
+        assertThat(stringResponse(response), containsString("isa has-genre"));
     }
 
     //TODO Prefix with Z to run last until TP Bug #13730 Fixed
@@ -475,7 +471,7 @@ public class GraqlControllerTest {
         assumeTrue(usingTitan());
 
         String fromId = graphContext.graph().getResourcesByValue("The Muppets").iterator().next().owner().getId().getValue();
-        String toId = graphContext.graph().getResourcesByValue("Kermit The Frog").iterator().next().owner().getId().getValue();
+        String toId = graphContext.graph().getResourcesByValue("comedy").iterator().next().owner().getId().getValue();
 
         String query = String.format("compute path from \"%s\" to \"%s\";", fromId, toId);
         Response response = sendGET(query, APPLICATION_HAL);
@@ -489,7 +485,7 @@ public class GraqlControllerTest {
         assumeTrue(usingTitan());
 
         String fromId = graphContext.graph().getResourcesByValue("The Muppets").iterator().next().owner().getId().getValue();
-        String toId = graphContext.graph().getResourcesByValue("Kermit The Frog").iterator().next().owner().getId().getValue();
+        String toId = graphContext.graph().getResourcesByValue("comedy").iterator().next().owner().getId().getValue();
 
         String query = String.format("compute path from \"%s\" to \"%s\";", fromId, toId);
         Response response = sendGET(query, APPLICATION_HAL);
@@ -503,7 +499,7 @@ public class GraqlControllerTest {
         assumeTrue(usingTitan());
 
         String fromId = graphContext.graph().getResourcesByValue("The Muppets").iterator().next().owner().getId().getValue();
-        String toId = graphContext.graph().getResourcesByValue("Kermit The Frog").iterator().next().owner().getId().getValue();
+        String toId = graphContext.graph().getResourcesByValue("comedy").iterator().next().owner().getId().getValue();
 
         String query = String.format("compute path from \"%s\" to \"%s\";", fromId, toId);
         Response response = sendGET(query, APPLICATION_HAL);
@@ -514,8 +510,8 @@ public class GraqlControllerTest {
     //TODO Prefix with Z to run last until TP Bug #13730 Fixed
     @Test
     public void ZGETGraqlComputePathWithHALTypeAndNoPath_ResponseIsEmptyJson(){
-        String fromId = graphContext.graph().getResourcesByValue("Marlon Brando").iterator().next().owner().getId().getValue();
-        String toId = graphContext.graph().getResourcesByValue("Kermit The Frog").iterator().next().owner().getId().getValue();
+        String fromId = graphContext.graph().getResourcesByValue("Apocalypse Now").iterator().next().owner().getId().getValue();
+        String toId = graphContext.graph().getResourcesByValue("comedy").iterator().next().owner().getId().getValue();
 
         String query = String.format("compute path from \"%s\" to \"%s\";", fromId, toId);
         Response response = sendGET(query, APPLICATION_HAL);
@@ -524,325 +520,8 @@ public class GraqlControllerTest {
         assertThat(jsonResponse(response), equalTo(Json.array()));
     }
 
-    @Test
-    public void POSTGraqlInsert_GraphCommitCalled(){
-        String query = "insert $x isa person;";
-
-        verify(mockGraph, times(0)).commit();
-
-        sendPOST(query);
-
-        verify(mockGraph, times(1)).commit();
-    }
-
-    @Test
-    public void POSTGraqlInsert_InsertWasExecutedOnGraph(){
-        doAnswer(answer -> {
-            graphContext.graph().commit();
-            return null;
-        }).when(mockGraph).commit();
-
-        String query = "insert $x isa person;";
-
-        int personCountBefore = graphContext.graph().getEntityType("person").instances().size();
-
-        sendPOST(query);
-
-        // refresh graph
-        graphContext.graph().close();
-
-        int personCountAfter = graphContext.graph().getEntityType("person").instances().size();
-
-        assertEquals(personCountBefore + 1, personCountAfter);
-    }
-
-    @Test
-    public void POSTMalformedGraqlQuery_ResponseStatusIs400(){
-        String query = "insert $x isa ;";
-        Response response = sendPOST(query);
-
-        assertThat(response.statusCode(), equalTo(400));
-    }
-
-    @Test
-    public void POSTMalformedGraqlQuery_ResponseExceptionContainsSyntaxError(){
-        String query = "insert $x isa ;";
-        Response response = sendPOST(query);
-
-        assertThat(exception(response), containsString("syntax error"));
-    }
-
-    @Test
-    public void POSTWithNoKeyspace_ResponseStatusIs400(){
-        String query = "insert $x isa person;";
-
-        Response response = RestAssured.with()
-                .body(query)
-                .post(String.format("http://%s:%s%s", HOST, PORT, REST.WebPath.Graph.GRAQL));
-
-        assertThat(response.statusCode(), equalTo(400));
-        assertThat(exception(response), containsString(MISSING_MANDATORY_REQUEST_PARAMETERS.getMessage(KEYSPACE)));
-    }
-
-    @Test
-    public void POSTWithNoQueryInBody_ResponseIs400(){
-        Response response = RestAssured.with()
-                .post(String.format("http://%s:%s%s", HOST, PORT, REST.WebPath.Graph.GRAQL));
-
-        assertThat(response.statusCode(), equalTo(400));
-        assertThat(exception(response), containsString(MISSING_REQUEST_BODY.getMessage()));
-    }
-
-    @Test
-    public void POSTGraqlMatch_ResponseStatusCodeIs405NotSupported(){
-        Response response = sendPOST("match $x isa name;");
-
-        assertThat(response.statusCode(), equalTo(405));
-    }
-
-    @Test
-    public void POSTGraqlMatch_ResponseExceptionContainsInsertOnlyAllowedMessage(){
-        Response response = sendPOST("match $x isa name;");
-
-        assertThat(exception(response), containsString("Only INSERT queries are allowed."));
-    }
-
-    @Test
-    public void POSTGraqlDelete_ResponseStatusCodeIs405NotSupported(){
-        Response response = sendPOST("match $x isa name; delete $x;");
-
-        assertThat(response.statusCode(), equalTo(405));
-    }
-
-    @Test
-    public void POSTGraqlDelete_ResponseExceptionContainsInsertOnlyAllowedMessage(){
-        Response response = sendPOST("match $x isa name; delete $x;");
-
-        assertThat(exception(response), containsString("Only INSERT queries are allowed."));
-    }
-
-    @Test
-    public void POSTGraqlCompute_ResponseStatusCodeIs405NotSupported(){
-        Response response = sendPOST("compute count in name;");
-
-        assertThat(response.statusCode(), equalTo(405));
-    }
-
-    @Test
-    public void POSTGraqlCompute_ResponseExceptionContainsInsertOnlyAllowedMessage(){
-        Response response = sendPOST("compute count in name;");
-
-        assertThat(exception(response), containsString("Only INSERT queries are allowed."));
-    }
-
-    @Test
-    public void POSTGraqlInsert_ResponseStatusIs200(){
-        String query = "insert $x isa person;";
-        Response response = sendPOST(query);
-
-        assertThat(response.statusCode(), equalTo(200));
-    }
-
-    @Test
-    public void POSTGraqlInsertNotValid_ResponseStatusCodeIs422(){
-        Response response = sendPOST("insert person plays movie;");
-
-        assertThat(response.statusCode(), equalTo(422));
-    }
-
-    @Test
-    public void POSTGraqlInsertNotValid_ResponseExceptionContainsValidationErrorMessage(){
-        Response response = sendPOST("insert person plays movie;");
-
-        assertThat(exception(response), containsString("is not of type"));
-    }
-
-    @Test
-    public void POSTGraqlInsert_ResponseContentTypeIsJson(){
-        Response response = sendPOST("insert $x isa movie;");
-
-        assertThat(response.contentType(), equalTo(APPLICATION_JSON));
-    }
-
-    @Test
-    public void POSTGraqlInsert_ResponseContainsOriginalQuery(){
-        String query = "insert $x isa person;";
-        Response response = sendPOST(query);
-
-        assertThat(originalQuery(response), equalTo(query));
-    }
-
-    @Test
-    public void POSTGraqlInsert_ResponseContainsCorrectNumberOfIds(){
-        String query = "insert $x isa person;";
-        Response response = sendPOST(query);
-
-        assertThat(1, equalTo(jsonResponse(response).asJsonList().size()));
-    }
-
-    @Test
-    public void POSTGraqlInsertWithOntology_GraphCommitIsCalled(){
-        String query = "insert thing sub entity;";
-
-        verify(mockGraph, times(0)).commit();
-
-        sendPOST(query);
-
-        verify(mockGraph, times(1)).commit();
-    }
-
-    @Test
-    public void DELETEGraqlDelete_GraphCommitCalled(){
-        String query = "match $x has title \"Heat\"; limit 1; delete $x;";
-
-        verify(mockGraph, times(0)).commit();
-
-        sendDELETE(query);
-
-        verify(mockGraph, times(1)).commit();
-    }
-
-    @Test
-    public void DELETEMalformedGraqlQuery_ResponseStatusIs400(){
-        String query = "match $x isa ; delete;";
-        Response response = sendDELETE(query);
-
-        assertThat(response.statusCode(), equalTo(400));
-    }
-
-    @Test
-    public void DELETEMalformedGraqlQuery_ResponseExceptionContainsSyntaxError(){
-        String query = "match $x isa ; delete;";
-        Response response = sendDELETE(query);
-
-        assertThat(exception(response), containsString("syntax error"));
-    }
-
-    @Test
-    public void DELETEWithNoKeyspace_ResponseStatusIs400(){
-        String query = "match $x isa person; delete;";
-
-        Response response = RestAssured.with()
-                .body(query)
-                .delete(String.format("http://%s:%s%s", HOST, PORT, REST.WebPath.Graph.GRAQL));
-
-        assertThat(response.statusCode(), equalTo(400));
-        assertThat(exception(response), containsString(MISSING_MANDATORY_REQUEST_PARAMETERS.getMessage(KEYSPACE)));
-    }
-
-    @Test
-    public void DELETEWithNoQueryInBody_ResponseIs400(){
-        Response response = RestAssured.with()
-                .delete(String.format("http://%s:%s%s", HOST, PORT, REST.WebPath.Graph.GRAQL));
-
-        assertThat(response.statusCode(), equalTo(400));
-        assertThat(exception(response), containsString(MISSING_REQUEST_BODY.getMessage()));
-    }
-
-    @Test
-    public void DELETEGraqlMatch_ResponseStatusCodeIs405NotSupported(){
-        Response response = sendDELETE("match $x isa name;");
-
-        assertThat(response.statusCode(), equalTo(405));
-    }
-
-    @Test
-    public void DELETEGraqlMatch_ResponseExceptionContainsDELETEOnlyAllowedMessage(){
-        Response response = sendDELETE("match $x isa name;");
-
-        assertThat(exception(response), containsString("Only DELETE queries are allowed."));
-    }
-
-    @Test
-    public void DELETEGraqlInsert_ResponseStatusCodeIs405NotSupported(){
-        Response response = sendDELETE("insert $x isa person;");
-
-        assertThat(response.statusCode(), equalTo(405));
-    }
-
-    @Test
-    public void DELETEGraqlInsert_ResponseExceptionContainsDELETEOnlyAllowedMessage(){
-        Response response = sendDELETE("insert $x isa person;");
-
-        assertThat(exception(response), containsString("Only DELETE queries are allowed."));
-    }
-
-    @Test
-    public void DELETEGraqlCompute_ResponseStatusCodeIs405NotSupported(){
-        Response response = sendDELETE("compute count in name;");
-
-        assertThat(response.statusCode(), equalTo(405));
-    }
-
-    @Test
-    public void DELETEGraqlCompute_ResponseExceptionContainsDELETEOnlyAllowedMessage(){
-        Response response = sendDELETE("compute count in name;");
-
-        assertThat(exception(response), containsString("Only DELETE queries are allowed."));
-    }
-
-    @Test
-    public void DELETEGraqlDelete_ResponseStatusIs200(){
-        String query = "match $x has name \"Robert De Niro\"; limit 1; delete $x;";
-        Response response = sendDELETE(query);
-
-        assertThat(response.statusCode(), equalTo(200));
-    }
-
-    @Test
-    public void DELETEGraqlDelete_DeleteWasExecutedOnGraph(){
-        doAnswer(answer -> {
-            graphContext.graph().commit();
-            return null;
-        }).when(mockGraph).commit();
-
-        String query = "match $x has name \"Martin Sheen\"; limit 1; delete $x;";
-
-        int personCountBefore = graphContext.graph().getEntityType("person").instances().size();
-
-        sendDELETE(query);
-
-        // refresh graph
-        graphContext.graph().close();
-
-        int personCountAfter = graphContext.graph().getEntityType("person").instances().size();
-
-        assertEquals(personCountBefore - 1, personCountAfter);
-    }
-
-    @Test
-    public void DELETEGraqlDeleteNotValid_ResponseStatusCodeIs422(){
-        // Not allowed to delete roles with incoming edges
-        Response response = sendDELETE("match $x label \"production-being-directed\"; delete $x;");
-
-        assertThat(response.statusCode(), equalTo(422));
-    }
-
-    @Test
-    public void DELETEGraqlDeleteNotValid_ResponseExceptionContainsValidationErrorMessage(){
-        // Not allowed to delete roles with incoming edges
-        Response response = sendDELETE("match $x label \"production-being-directed\"; delete $x;");
-
-        assertThat(exception(response), containsString("cannot be deleted"));
-    }
-
-    @Test
-    public void DELETEGraqlDelete_ResponseContentTypeIsJson(){
-        Response response = sendDELETE("match $x has name \"Harry\"; limit 1; delete $x;");
-
-        assertThat(response.contentType(), equalTo(APPLICATION_JSON));
-    }
-
-    @Test
-    public void DELETEGraqlDelete_ResponseContainsOriginalQuery(){
-        String query = "match $x has name \"Miranda Heart\"; limit 1; delete $x;";
-        Response response = sendDELETE(query);
-
-        assertThat(originalQuery(response), equalTo(query));
-    }
-
     private Response sendGET(String acceptType){
-        return sendGET("match $x isa person;", acceptType, false, false, -1);
+        return sendGET("match $x isa movie;", acceptType, false, false, -1);
     }
 
     private Response sendGET(String match, String acceptType){
@@ -859,20 +538,6 @@ public class GraqlControllerTest {
                 .queryParam(LIMIT_EMBEDDED, limitEmbedded)
                 .accept(acceptType)
                 .get(String.format("http://%s:%s%s", HOST, PORT, REST.WebPath.Graph.GRAQL));
-    }
-
-    private Response sendPOST(String query){
-        return RestAssured.with()
-                .queryParam(KEYSPACE, mockGraph.getKeyspace())
-                .body(query)
-                .post(String.format("http://%s:%s%s", HOST, PORT, REST.WebPath.Graph.GRAQL));
-    }
-
-    private Response sendDELETE(String query){
-        return RestAssured.with()
-                .queryParam(KEYSPACE, mockGraph.getKeyspace())
-                .body(query)
-                .delete(String.format("http://%s:%s%s", HOST, PORT, REST.WebPath.Graph.GRAQL));
     }
 
     protected static String exception(Response response){
